backend/requirements.txt
# OpenAI and AI Dependencies
openai>=1.0.0
pydantic-ai>=0.0.10
pydantic>=2.0.0
pydantic-core>=2.0.0

# Supabase
supabase>=2.0.0
postgrest>=0.10.0

# Utilities
python-dotenv>=1.0.0
pyjwt>=2.8.0
requests>=2.31.0

# Image Processing (optional - for future image handling)
# pillow>=10.0.0
